backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
google-genai>=1.0.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
mcp>=1.22.0
pydantic>=2.0.0
websockets>=12.0
